NUP-2368 Moved Link unit tests to a source dedicated file LinkTest.cpp. NUP-2368 Updated description of Link::shiftBufferedData as recommended by Scott. NUP-2368 Clarified intended propagation of data depending on the propagationDelay arg in the Link constructor documentation.
